ledger-api-test-tool: move into //ledger/ledger-api-test-tool.
